If the first TAP tape header is bad, use the second
This is the feature making the Commodore tape format so reliable, and so
slow. In reality, the backup will seldom be used here because the TAP
reading code will abort as soon as a byte with bad parity is found. The
test file test2.tap has been corrupted with a bad checksum but good
parity to test this recovery path.