Disable the Adélie CI build
The Docker image is no longer able to be updated when run due to a
"Permission denied" error. The symptoms match that of Adélie servers
blocking GitHub Actions from access.

This eliminates big-endian tests, and although it is becoming
exceedingly rare these days, it's something that kind of retrocomputing
enthusiast that uses fvcbm may very well be relying on.